Addn train and val steps
